# openapi.yaml
openapi: 3.1.0
info:
  title: Tinia public API
  version: 0.0.1-SNAPSHOT
  description: API to interact with Tinia services.

servers:
  - url: http://localhost:8080/api/v1

paths:
  /upload:
    post:
      operationId: uploadPlan
      summary: Upload a file containing a medication plan
      description:
        tbd
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: The PDF or image file to upload.
                processOcr:
                  type: boolean
                  description: Whether to process the uploaded file via OCR.
                  default: true
              required:
                - file
      responses:
        '202':
          description: File accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input

  /upload/jobs/{jobId}/pages/{pageNumber}/continue:
    post:
      operationId: continueProcessing
      summary: Continue the processing after user validated current findings
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pageNumber
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContinueTask'
      responses:
        '200':
          description: "ok"

  /upload/jobs/{jobId}/pages/{pageNumber}/image:
    get:
      operationId: getImage
      summary: Retrieve a rendered page image
      description: Returns the rendered image of a page belonging to a processing job.
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pageNumber
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Rendered page image
          content:
            image/png:
              schema:
                type: string
                format: binary
        '404':
          description: Job or page not found
        '410':
          description: Job results no longer available
components:
  schemas:
    Task:
      type: object
      properties:
        jobId:
          type: string
          format: uuid
      required:
        - jobId
    ContinueTask:
      type: object
      properties:
        cells:
          type: array
          items:
            $ref: '#/components/schemas/Cell'
        template:
          $ref: '#/components/schemas/template'
      required:
        - cells
        - template
    Cell:
      type: object
      properties:
        column:
          type: integer
        row:
          type: integer
        width:
          type: integer
        height:
          type: integer
        x:
          type: integer
        y:
          type: integer
        cellType:
          type: string
          enum:
            - HEADER
            - DATA
        id:
          type: string
          format: uuid
      required:
        - column
        - row
        - width
        - height
        - x
        - y
        - id
    template:
      type: object
      properties:
        author:
          type: string
        bmpMappings:
          type: array
          items:
            type: object
            properties:
              bmpField:
                type: string
              headerCellIds:
                type: array
                items:
                  type: string
                  format: uuid
        createdAt:
          type: string
        customFields:
          type: array
          items:
            type: string
        notes:
          type: array
          items:
            type: string
        templateName:
          type: string
        version:
          type: string
      required:
        - bmpMappings